[ETFE-1568] Refactor to use consistent TraderModel
