+++ app/models/response/emcsTfe/GetMovementResponse.scala
@@ -17,7 +17,7 @@
 package models.response.emcsTfe
 
 import models.DestinationType
-import models.submitExplainDelay.TraderModel
+import models.common.TraderModel
 import play.api.libs.json.{Json, Reads}
 
 import java.time.LocalDate
@@ -30,7 +30,7 @@ case class GetMovementResponse(arc: String,
                                deliveryPlaceTrader: Option[TraderModel],
                                localReferenceNumber: String,
                                eadStatus: String,
-                               consignorTrader: ConsignorTraderModel,
+                               consignorTrader: TraderModel,
                                dateOfDispatch: LocalDate,
                                journeyTime: String,
                                numberOfItems: Int) {
